+++ payment/src/test/java/com/ll/payment/deposit/DepositServiceFailedUnitTest.java
@@ -4,18 +4,9 @@ import com.fasterxml.uuid.Generators;
 import com.ll.payment.deposit.model.entity.Deposit;
 import com.ll.payment.deposit.model.enums.DepositHistoryType;
 import com.ll.payment.deposit.model.enums.DepositStatus;
-import com.ll.payment.deposit.model.exception.DepositNotFoundException;
-import com.ll.payment.deposit.model.exception.DepositAlreadyExistsException;
-import com.ll.payment.deposit.model.exception.DuplicateDepositTransactionException;
-import com.ll.payment.deposit.model.exception.DepositBalanceNotEmptyException;
-import com.ll.payment.deposit.model.exception.InvalidDepositStatusTransitionException;
-import com.ll.payment.deposit.model.exception.InsufficientDepositBalanceException;
-import com.ll.payment.deposit.model.exception.RefundTargetNotFoundException;
+import com.ll.payment.deposit.model.exception.*;
 import com.ll.payment.deposit.model.vo.request.DepositDeleteRequest;
 import com.ll.payment.deposit.model.vo.request.DepositTransactionRequest;
-import com.ll.payment.deposit.model.vo.response.DepositDeleteResponse;
-import com.ll.payment.deposit.model.vo.response.DepositResponse;
-import com.ll.payment.deposit.model.vo.response.DepositTransactionResponse;
 import com.ll.payment.deposit.repository.DepositRepository;
 import com.ll.payment.deposit.service.DepositHistoryService;
 import com.ll.payment.deposit.service.DepositServiceImpl;
@@ -38,7 +29,7 @@ import static org.mockito.Mockito.*;
 @SuppressWarnings({"NonAsciiCharacters", "FieldCanBeLocal"})
 @ExtendWith(MockitoExtension.class)
 @DisplayName("DepositService 단위 테스트")
-public class DepositServiceUnitTest {
+public class DepositServiceFailedUnitTest {
 
     @Mock
     private DepositRepository depositRepository;
@@ -57,9 +48,7 @@ public class DepositServiceUnitTest {
 
     private Deposit deposit;
     private DepositTransactionRequest transactionRequest;
-    private DepositTransactionResponse transactionResponse;
     private DepositDeleteRequest deleteRequest;
-    private DepositDeleteResponse deleteResponse;
 
     @BeforeEach
     void setUp() {
@@ -82,13 +71,6 @@ public class DepositServiceUnitTest {
                 .thenReturn(Optional.of(deposit));
     }
 
-    private void verifySuccessFlow() {
-        verify(depositRepository).findByUserCode(USER_CODE);
-        verify(depositHistoryService).validateDuplicate(REF_CODE);
-        verify(depositHistoryService).saveSuccessHistory(any());
-        verify(depositHistoryService, never()).saveFailedHistory(any(), any(), any(), any());
-    }
-
     private void verifyFailedFlow(DepositHistoryType failedType) {
         verify(depositRepository).findByUserCode(USER_CODE);
         verify(depositHistoryService).validateDuplicate(REF_CODE);
@@ -99,13 +81,6 @@ public class DepositServiceUnitTest {
                 eq(failedType),
                 any(Exception.class)
         );
-    }
-
-    private void verifyRefundSuccessFlow() {
-        verify(depositRepository).findByUserCode(USER_CODE);
-        verify(depositHistoryService).validateDuplicateForRefund(REF_CODE);
-        verify(depositHistoryService).saveSuccessHistory(any());
-        verify(depositHistoryService, never()).saveFailedHistory(any(), any(), any(), any());
     }
 
     private void verifyRefundFailedFlow() {
@@ -120,77 +95,10 @@ public class DepositServiceUnitTest {
         );
     }
 
-    private void assertTransactionResponse(Long balance, DepositHistoryType charge) {
-        assertNotNull(transactionResponse);
-        assertEquals(deposit.getBalance(), balance);
-        assertEquals(deposit.getCode(), transactionResponse.depositCode());
-        assertEquals(AMOUNT, transactionResponse.amount());
-        assertEquals(balance, transactionResponse.balanceAfter());
-        assertEquals(charge, transactionResponse.historyType());
-        assertEquals(REF_CODE, transactionResponse.referenceCode());
-    }
-
-    private void assertDeleteResponse() {
-        verify(depositRepository).findByUserCode(USER_CODE);
-        verify(depositRepository).save(argThat(d -> d.getDepositStatus() == DepositStatus.CLOSED));
-        assertNotNull(deleteResponse);
-        assertEquals(USER_CODE, deleteResponse.userCode());
-        assertEquals(DELETE_REASON, deleteResponse.closedReason());
-        assertEquals(DepositStatus.CLOSED, deposit.getDepositStatus());
-        assertEquals(DepositStatus.CLOSED, deleteResponse.depositStatus());
-    }
-
     /* ==========================
         Deposit 생성 및 삭제 Tests
        ========================== */
 
-    // 성공 케이스
-    @Test
-    void 기존_Deposit_이_존재하지_않을_때_신규_Deposit_을_생성() {
-        // given
-        mockDepositNotFound();
-        when(depositRepository.save(any())).thenAnswer(inv -> inv.getArgument(0));
-
-        // when
-        DepositResponse response = depositService.createDeposit(USER_CODE);
-
-        // then
-        verify(depositRepository).findByUserCode(USER_CODE);
-        verify(depositRepository).save(argThat(d -> d.getDepositStatus() == DepositStatus.ACTIVE));
-        assertNotNull(response);
-        assertEquals(USER_CODE, response.userCode());
-    }
-
-    @Test
-    void 기존_Deposit_이_Close_상태로_존재할_때_해당_Deposit_상태를_Active_로_변경() {
-        // given
-        deposit.setClosed();
-        mockDepositFound();
-
-        // when
-        DepositResponse response = depositService.createDeposit(USER_CODE);
-
-        // then
-        verify(depositRepository).findByUserCode(USER_CODE);
-        assertNotNull(response);
-        assertEquals(USER_CODE, response.userCode());
-        assertEquals(DepositStatus.ACTIVE, deposit.getDepositStatus());
-    }
-
-    @Test
-    void Deposit_의_상태를_Close_로_변경() {
-        // given
-        mockDepositFound();
-        when(depositRepository.save(any())).thenAnswer(inv -> inv.getArgument(0));
-
-        // when
-        deleteResponse = depositService.deleteDepositByUserCode(USER_CODE, deleteRequest);
-
-        // then
-        assertDeleteResponse();
-    }
-
-    // 실패 케이스
     @Test
     void 기존_Deposit_이_Active_상태로_존재할_때_신규_Deposit_생성_시도_시_예외_발생() {
         // given
@@ -259,23 +167,6 @@ public class DepositServiceUnitTest {
         Deposit 충전 Tests
        =================== */
 
-    // 성공 케이스
-    @Test
-    void Deposit_충전_성공() {
-        // given
-        mockDepositFound();
-        doNothing().when(depositHistoryService).validateDuplicate(REF_CODE);
-        doNothing().when(depositHistoryService).saveSuccessHistory(any());
-
-        // when
-        transactionResponse = depositService.chargeDeposit(USER_CODE, transactionRequest);
-
-        // then
-        verifySuccessFlow();
-        assertTransactionResponse(AMOUNT, DepositHistoryType.CHARGE);
-    }
-
-    // 실패 케이스
     @Test
     void Deposit_충전_실패_이미_존재하는_referenceCode() {
         // given
@@ -330,23 +221,6 @@ public class DepositServiceUnitTest {
         Deposit 출금 Tests
        =================== */
 
-    // 성공 케이스
-    @Test
-    void Deposit_출금_성공() {
-        // given
-        deposit.charge(AMOUNT, REF_CODE);
-        mockDepositFound();
-        doNothing().when(depositHistoryService).validateDuplicate(REF_CODE);
-        doNothing().when(depositHistoryService).saveSuccessHistory(any());
-        // when
-        transactionResponse = depositService.withdrawDeposit(USER_CODE, transactionRequest);
-
-        // then
-        verifySuccessFlow();
-        assertTransactionResponse(0L, DepositHistoryType.WITHDRAW);
-    }
-
-    // 실패 케이스
     @Test
     void Deposit_출금_실패_잔액_부족() {
         // given
@@ -418,24 +292,6 @@ public class DepositServiceUnitTest {
         Deposit 결제 Tests
        =================== */
 
-    // 성공 케이스
-    @Test
-    void Deposit_결제_성공() {
-        // given
-        deposit.charge(AMOUNT, REF_CODE);
-        mockDepositFound();
-        doNothing().when(depositHistoryService).validateDuplicate(REF_CODE);
-        doNothing().when(depositHistoryService).saveSuccessHistory(any());
-
-        // when
-        transactionResponse = depositService.paymentDeposit(USER_CODE, transactionRequest);
-
-        // then
-        verifySuccessFlow();
-        assertTransactionResponse(0L, DepositHistoryType.PAYMENT);
-    }
-
-    // 실패 케이스
     @Test
     void Deposit_결제_실패_잔액_부족() {
         // given
@@ -506,22 +362,6 @@ public class DepositServiceUnitTest {
         Deposit 환불 Tests
        =================== */
 
-    // 성공 케이스
-    @Test
-    void Deposit_환불_성공() {
-        // given
-        mockDepositFound();
-        doNothing().when(depositHistoryService).validateDuplicateForRefund(REF_CODE);
-        doNothing().when(depositHistoryService).saveSuccessHistory(any());
-        // when
-        transactionResponse = depositService.refundDeposit(USER_CODE, transactionRequest);
-
-        // then
-        verifyRefundSuccessFlow();
-        assertTransactionResponse(AMOUNT, DepositHistoryType.REFUND);
-    }
-
-    // 실패 케이스
     @ParameterizedTest
     @ValueSource(classes = {
             DuplicateDepositTransactionException.class,
